Label-Microservice/requirements.txt
-i https://pypi.org/simple
argo-models==2.2.1a0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blis==0.2.4
boto3==1.9.199
botocore==1.12.199
Bottleneck==1.2.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
cytoolz==0.10.0
dataclasses==0.6
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.0
docker==4.0.2
docutils==0.14
entrypoints==0.3
-e git+https://github.com/kubeflow/fairing.git#egg=fairing
fastai==1.0.55
fastprogress==0.1.21
fire==0.1.3
future==0.17.1
google-api-core==1.14.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.17.0
google-cloud-core==1.0.3
google-cloud-pubsub==0.45.0
google-cloud-storage==1.17.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.22.0
html5lib==1.0.1
httplib2==0.13.1
idna==2.8
importlib-metadata==0.19
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.1
jellyfish==0.6.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kfp==0.1.25
kfp-server-api==0.1.18.3
kiwisolver==1.1.0
kubernetes==9.0.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mdparse==0.13
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numexpr==2.6.9
numpy==1.16.4
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.0.2
packaging==19.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
passlib==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
pluggy==0.12.0
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.6
pyasn1-modules==0.2.5
pycparser==2.19
pyemd==0.5.1
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
Pyphen==0.9.5
pyrsistent==0.15.3
pytest==5.0.1
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.1
PyYAML==5.1.2
pyzmq==18.0.2
qtconsole==4.5.1
regex==2019.6.8
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
spacy==2.1.6
srsly==0.0.7
tabulate==0.8.3
terminado==0.8.2
testpath==0.4.2
textacy==0.7.0
thinc==7.0.8
timeout-decorator==0.4.1
toolz==0.10.0
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
typing==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
widgetsnbextension==3.5.0
zipp==0.5.2
